Updated DC-DC convertor, fixed CH340G Vout
I changed switching regulator to MP2259, because MP2359 is not recommended for new designs and is therefore more futureproof.
Resistor values was also changed to lower the voltage from 3.5V to 3.33V.
I also changed USB-C type to TYPE-C-31-M-12, because I have it :) It also seems to be more popular and I saw many compatible clones on Aliexpress for example.
CH340G was set to operate with 5V signals, which isn't good for ESP, so I changed it. I also changed the part to CH340C to avoid using external crystal (which will be optional for CH340G compatibility).